extract suggestions logic to its own service SuggestionService.kt
